Move all atoms to internal modules
