+++ lib/provider/ui/windows/solitaire/foudation-atom.ts
