+++ lib/provider/ui/windows/solitaire/game.tsx
@@ -15,12 +15,12 @@ import {
   cardWidth,
 } from "~/lib/solitaire-constant";
 import { stock } from "~/lib/solitaire-global";
-import { tableauAtom } from "~/lib/solitaire-tableau-atom";
 import { isBlack, shuffleCards, updateCardInfo } from "~/lib/solitaire-utils";
 
 import { Foundation } from "./foundation";
 import { Pile } from "./pile";
 import { Tableau } from "./tableau";
+import { tableauAtom } from "./tableau-atom";
 import { Waste } from "./waste";
 
 const Wrapper = styled.div`
