+++ lib/provider/ui/windows/solitaire/holder.tsx
@@ -8,8 +8,9 @@ import { useDrop } from "react-dnd";
 import styled from "styled-components";
 
 import { cardHeight, cardWidth } from "~/lib/solitaire-constant";
-import { moveCardAtom } from "~/lib/solitaire-move-card-atom";
 import img from "~/lib/solitaire-spritesheet.png";
+
+import { moveCardAtom } from "./move-card-atom";
 
 const HolderStyled = styled.div`
   width: ${cardWidth}px;
