+++ lib/provider/ui/windows/solitaire/tableau-atom.ts
