+++ lib/provider/ui/windows/solitaire/waste-atom.ts
